+++ backend/sivujetti/tests/src/Page/UpsertPageBlockStylesTest.php
@@ -73,7 +73,7 @@ final class UpsertPageBlockStylesTest extends PagesControllerTestCase {
     }
     private function verifyOverwroteGeneratedCssFile(\TestState $state): void {
         $expectedBlockStylePart = $this->cssGenTestUtils->generateCachedBlockStyles($state->inputData->styles);
-        $actual = $this->cssGenTestUtils->getActualGeneratedCss();
+        $actual = $this->cssGenTestUtils->getActualGeneratedCss(ignore: "blockTypeBaseCss");
         $expected = $this->cssGenTestUtils->generateExpectedGeneratedCssContent(expectedBlockStyles: $expectedBlockStylePart);
         $this->assertEquals($expected, $actual);
     }