+++ e2e-tests/tests/block-styles-edit-test.js
@@ -15,7 +15,7 @@ describe('Block', () => {
     it('User can edit styles', browser => {
         const testBlock = testPageData.blocks.find(({type}) => type === 'Paragraph');
         const testBlockBlockTreeBtnLocator = `.block-tree li[data-block-id="${testBlock.id}"] button`;
-        const stylesTabBtnLocator = `${vars.inspectorPanelLocator} > ul .tab-item:nth-of-type(2)`;
+        const stylesTabBtnLocator = `${vars.inspectorPanelLocator} > ul .tab-item:nth-child(2)`;
         const stylesTextareaLocator = `${vars.inspectorPanelLocator} > div:last-of-type textarea`;
         const validationErrorLocator = `${stylesTextareaLocator} + *`;
         const testBlockInPageLocator = `body [data-block="${testBlock.id}"]`;
@@ -34,8 +34,8 @@ describe('Block', () => {
             .assert.not.elementPresent(validationErrorLocator)
             //
             .frame(0)
-            .execute(testBLockLocator =>
-                window.getComputedStyle(document.querySelector(testBLockLocator)).color
+            .execute(testBlockLocator =>
+                window.getComputedStyle(document.querySelector(testBlockLocator)).color
             , [testBlockInPageLocator], result => {
                 browser.assert.equal(result.value, 'rgb(255, 0, 0)');
             })
@@ -47,8 +47,8 @@ describe('Block', () => {
             //
             .navigateTo(env.makeUrl(testPageData.slug))
             .waitForElementPresent(testBlockInPageLocator)
-            .execute(testBLockLocator =>
-                window.getComputedStyle(document.querySelector(testBLockLocator)).color
+            .execute(testBlockLocator =>
+                window.getComputedStyle(document.querySelector(testBlockLocator)).color
             , [testBlockInPageLocator], result => {
                 browser.assert.equal(result.value, 'rgb(255, 0, 0)');
             });