+++ changelog.md
@@ -8,10 +8,10 @@
 - An ability to drag blocks between block tree's branches
 
 ## Fixed
-- The page doesn't blow up in edit mode anymore, if some of the blocks contain some unusual css
+- An issue, where the page blew up in edit mode, if some of the blocks contained some unusual css
 
 ## Changed
-- .
+- The event type (onChange -> onInput) that refreshes styles preview in edit mode
 
 # 0.9.0 (2022-04-08)